requirements.txt
streamlit>=1.28
streamlit-chat
langchain>=0.1.0
openai>=0.28
streamlit-feedback
azure-identity==1.15.0
azure-ai-documentintelligence==1.0.0b3
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0
python-docx==1.1.0
PyPDF2==3.0.1
PyMuPDF==1.23.16
langchain-community==0.0.10
langchain-core==0.1.8
Pillow==10.2.0
tabula-py==2.9.0
tiktoken==0.5.2
markdown
matplotlib
seaborn
tiktoken
aiofiles 
asyncio